Remove new files as they didn't work as expected
