+++ src/modules/createlistelement.js
@@ -7,7 +7,7 @@ const listElement = {
 const createlistelement = (str, arraylist) => {
   const newlistelement = Object.create(listElement);
   newlistelement.description = str;
-  newlistelement.index = arraylist.length+1;
+  newlistelement.index = arraylist.length + 1;
   arraylist.push(newlistelement);
 };
